Add script for setting up RCP branch
I copied the 20C script and then made what I think are the necessary
modifications to continue the run from 2006 - 2025. It'll be 4 5-year segments,
which I think will need ~26 walltime hours.

Note that it looks to get restart files from archive, but if nothing is
available then we assume the restarts were moved to campaign before scrubbing
the archive. I should move the 20th C restarts to this project's campaign
storage and update the 20thC_portion script