Transfer logs and restart files as well
This commit actually enables the globus transfer instead of printing out
diagnostic information. It copies every time series file (grouped by component
and frequency, so all monthly ocean files are in one transfer and daily ocean
files are in another, 6-hourly atmosphere are in a third, and so one) as well
as every log file (in a single transfer) and the 2006 or 2026 restart files
(also in a single transfer per restart year).

I made some small tweaks to how the transfers are labeled and saved the globus
transfer lists in a new subdirectory globus_lists/