Fixed issue in gen_file_list()
gen_file_list() was written in such a way that it wasn't properly transfering
the pop.h.nday1 variables (SRC_ROOT was tied to the pop.h.ecosys.nday1 stream).
The fix was to remove ${stream} from the definition of SRC_ROOT and include the
full stream path in the transfer list.

I also added the option to specify an ensemble member as a command-line
argument.